+++ src/Paysera/WalletApi/Container.php
@@ -97,6 +97,7 @@ class Paysera_WalletApi_Container
 
         if (count($parameters) > 0) {
             $dispatcher->addSubscriber(new Paysera_WalletApi_Listener_ParameterSetter($parameters));
+            $dispatcher->addSubscriber(new Paysera_WalletApi_Listener_PayseraUserIpSetter($parameters));
         }
 
         $dispatcher->addSubscriber($requestSigner);